refactor(sal): the group header names its kind as an ordinal and its zone start as a byte, the writer owns the open zone and its LSN, deferred publication is a guard over a cursor, the reader hands the worker its own slot, and the eight engine-only bits leave gnitz-wire
`SalMessageKind` is a `wire_enum!` with `ZoneCommit` as the sentinel's own
kind; `ZoneMark` is gone. `DirectGroup` carries kind, lsn and `zoned`, and
`SalWriter::write` is the one write primitive: it opens the zone on the first
zoned group it admits, `close_zone` writes the sentinel from the LSN it holds,
and `DeferredPublication` publishes on drop by re-walking the laid-out headers.
`SalReader::next` yields `(msg, slot)` and steps over groups without a slot,
so the worker decodes once at the top of `dispatch`. The holdback is a function
of the kind alone; `fit_relay`/`past_reclaim_margin` replace the exposed
capacity arithmetic; the tail-wide `first_group_slots` gives way to the
per-group re-slice count in the worker's replay line.

+++ crates/gnitz-server/src/runtime/orchestration/master/tests/fixtures.rs
@@ -50,7 +50,7 @@ pub(super) fn test_dispatcher(worker_pids: Vec<i32>, catalog: *mut CatalogEngine
         worker_pids,
         catalog,
         0,
-        SalWriter::new(sal, -1, SAL_SIZE as u64, nw),
+        SalWriter::new(sal, -1, SAL_SIZE, nw),
         Rc::new(W2mReceiver::new(rings)),
         vec![-1; nw],
     )
